PHPCS/PHPCBF: small follow-up tweak
Follow up on 143

In contrast to earlier information, arrangements are now being made to allow the package to continue under its original name on Packagist. The commit (in the new repo) to rename the package [has been reverted](https://github.com/PHPCSStandards/PHP_CodeSniffer/pull/135).

Sorry for the confusion and thank you for understanding.